Move things out of example
